django_ganglioside/requirements/production.txt
aiofiles==23.2.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.1
black==23.11.0
blinker==1.9.0
celery==5.3.4
certifi==2025.8.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.3.0
fastapi==0.104.1
flake8==6.1.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==0.2.1
kiwisolver==1.4.7
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.8.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==1.24.3
packaging==25.0
pandas==2.1.3
pathspec==0.12.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==5.17.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyflakes==3.1.0
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
rpds-py==0.27.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.27.0
statsmodels==0.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
# Production WSGI Server
gunicorn==21.2.0

# WhiteNoise for static file serving
whitenoise==6.6.0

# Error tracking
sentry-sdk==1.40.0

# Security
django-ratelimit==4.1.0
